chore: MemeTypeButtonSection 공용컴포넌트로 분리
